Export useAuth to new file and move useAuth and AuthContext to new folder
