+++ frontend/src/components/Common/Navbar/Navbar.tsx
@@ -1,7 +1,7 @@
 import { Link } from "react-router-dom";
 import NavLink from "./NavLink";
 import { useState } from "react";
-import { useAuth } from "../../Context/AuthContext";
+import { useAuth } from "../../Context/Auth/useAuth";
 
 function Navbar() {
   const [isMenuOpen, setIsMenuOpen] = useState(false);
