+++ frontend/src/components/NotAuthenticated/RegisterForm.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import RegisterForm from "./RegisterForm";
-import { AuthProvider } from "../Context/AuthContext";
+import { AuthProvider } from "../Context/Auth/AuthContext";
 import { BrowserRouter as Router } from "react-router-dom";
 
 describe("RegisterForm Component", () => {
