+++ frontend/src/components/NotAuthenticated/RegisterForm.tsx
@@ -4,7 +4,7 @@ import { useSpring, animated } from "react-spring";
 import { ClipLoader } from "react-spinners";
 import api from "../../services/api.service";
 import { useNavigate } from "react-router-dom";
-import { useAuth } from "../Context/AuthContext";
+import { useAuth } from "../Context/Auth/useAuth";
 import isValidEmail from "../../utils/email.utils";
 
 function RegisterForm() {
