+++ frontend/src/pages/Authenticated/TodoListsPage.test.tsx
@@ -1,6 +1,6 @@
 import TodoListsPage from "./TodoListsPage";
 import { render, screen } from "@testing-library/react";
-import { AuthProvider } from "../../components/Context/AuthContext";
+import { AuthProvider } from "../../components/Context/Auth/AuthContext";
 import { BrowserRouter as Router } from "react-router-dom";
 
 describe("TodoListsPage Component", () => {
