+++ frontend/src/pages/Authenticated/TodoListsPage.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import { TextField } from "@mui/material";
 import TodoList from "../../components/Authenticated/TodoList";
-import { useAuth } from "../../components/Context/AuthContext";
+import { useAuth } from "../../components/Context/Auth/useAuth";
 import { newList, getAllLists, deleteList } from "../../services/list.service";
 
 interface NewListFormProps {
